Reorganize python scripts folder
