+++ scripts/src/utils/file_utils.py
@@ -1,4 +1,5 @@
 import os
+from pathlib import Path
 from typing import List
 
 
@@ -14,3 +15,7 @@ def list_dirs(path: str) -> List[str]:
         if os.path.isdir(path_to_file):
             dirs.append(file)
     return dirs
+
+
+def get_absolute_path(path: str) -> str:
+    return str(Path(path).resolve())
