+++ scripts/src/utils/runners/abstract_task_runner.py
@@ -2,18 +2,19 @@ from abc import ABC, abstractmethod
 from pathlib import Path
 from typing import Any, Dict, Tuple
 
-from utils.models.cli_models import TaskFlagArgs, TaskNamedArgs
-from utils.run_process_utils import run_in_subprocess
+from src.utils.models.cli_models import TaskFlagArgs, TaskNamedArgs
+from src.utils.run_process_utils import run_in_subprocess
 
 
 def get_common_named_arguments(
         step_id,
         script_arguments,
+        input_file_name_in_kwargs: str,
         **kwargs,
 ) -> Dict[TaskNamedArgs, str]:
     return {
         TaskNamedArgs.INPUT_FILE:
-            kwargs['build_initial_graph_filename'](step_id, script_arguments),
+            kwargs[input_file_name_in_kwargs](step_id, script_arguments),
         TaskNamedArgs.OUTPUT_PATH:
             kwargs['build_output_dir_name'](step_id, script_arguments),
         TaskNamedArgs.LANGUAGE: script_arguments.language,
@@ -28,7 +29,7 @@ class AbstractTaskRunner(ABC):
     task_prefix = ':code-submissions-clustering-plugin:'
 
     # The root folder of the initial project
-    PROJECT_DIR = Path(__file__).parent.parent.parent.parent
+    PROJECT_DIR = Path(__file__).parent.parent.parent.parent.parent
 
     @abstractmethod
     def build_arguments(self, *args, **kwargs) \
@@ -54,8 +55,6 @@ class AbstractTaskRunner(ABC):
         """Run task and return process stderr."""
         named_args, flag_args = self.build_arguments(*args, **kwargs)
         cmd = self.configure_cmd(named_args, flag_args)
-
-        print(cmd)
         return_code, stdout = run_in_subprocess(cmd.split(), self.PROJECT_DIR)
         if return_code != 0:
             return ''
