+++ scripts/src/utils/runners/calculate_dist_runner.py
@@ -1,7 +1,7 @@
 from typing import Any, Dict, Tuple
 
-from utils.models.cli_models import TaskFlagArgs, TaskNamedArgs
-from utils.runners.abstract_task_runner import AbstractTaskRunner, get_common_named_arguments
+from src.utils.models.cli_models import TaskFlagArgs, TaskNamedArgs
+from src.utils.runners.abstract_task_runner import AbstractTaskRunner, get_common_named_arguments
 
 
 class CalculateDistRunner(AbstractTaskRunner):
@@ -16,5 +16,5 @@ class CalculateDistRunner(AbstractTaskRunner):
             **kwargs,
     ) -> Tuple[Dict[TaskNamedArgs, Any], Dict[TaskFlagArgs, bool]]:
         """Build arguments for 'calculate-dist' CLI."""
-        named_args = get_common_named_arguments(step_id, script_arguments, **kwargs)
+        named_args = get_common_named_arguments(step_id, script_arguments, 'build_initial_graph_filename', **kwargs)
         return named_args, TaskFlagArgs.get_all_flags(script_arguments)
