+++ scripts/src/utils/runners/load_runner.py
@@ -1,8 +1,8 @@
 from argparse import Namespace
 from typing import Any, Dict, Tuple
 
-from utils.models.cli_models import TaskFlagArgs, TaskNamedArgs
-from utils.runners.abstract_task_runner import AbstractTaskRunner, get_common_named_arguments
+from src.utils.models.cli_models import TaskFlagArgs, TaskNamedArgs
+from src.utils.runners.abstract_task_runner import AbstractTaskRunner, get_common_named_arguments
 
 
 class LoadRunner(AbstractTaskRunner):
@@ -17,5 +17,5 @@ class LoadRunner(AbstractTaskRunner):
             **kwargs,
     ) -> Tuple[Dict[TaskNamedArgs, Any], Dict[TaskFlagArgs, bool]]:
         """Build arguments for 'load' CLI."""
-        named_args = get_common_named_arguments(step_id, script_arguments, **kwargs)
+        named_args = get_common_named_arguments(step_id, script_arguments, 'build_solutions_file_name', **kwargs)
         return named_args, TaskFlagArgs.get_all_flags(script_arguments)
